instance GoogleUtils and break up the code into separate files
