+++ compat/google/GoogleExtensions.cs
@@ -1,0 +1,11 @@
+﻿using Google.Apis.Calendar.v3.Data;
+
+namespace d9.utl.compat.google;
+public static class GoogleExtensions
+{
+    /// <summary>
+    /// Converts a normal <see cref="DateTime"/> to a Google Calendar <see cref="EventDateTime"/>.</summary>
+    /// <param name="dateTime">The <see cref="DateTime"/> to convert.</param>
+    /// <returns>An <see cref="EventDateTime"/> corresponding to the given <c><paramref name="dateTime"/></c>.</returns>
+    public static EventDateTime ToEventDateTime(this DateTime dateTime) => new() { DateTime = dateTime };
+}
